requirements.txt
# OSS SDK
oss2>=2.18.0

# CLI框架
click>=8.1.0

# 配置文件处理
PyYAML>=6.0

# 进度条
tqdm>=4.65.0

# 日志增强
colorlog>=6.7.0

# 图片处理（可选）
Pillow>=10.0.0

# 表格格式化
tabulate>=0.9.0

# 剪贴板支持（可选）
pyperclip>=1.8.0

# 开发依赖
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
